+++ package.json
@@ -41,15 +41,13 @@
     "@jaenjs/jaen": "3.0.0-beta.20",
     "@react-icons/all-files": "^4.1.0",
     "@sendgrid/mail": "^7.6.2",
-    "@shopify/shopify-api": "^2.0.0",
     "@snek-at/react-lottie": "^1.0.0-beta.3",
-    "@snek-at/storybook-addon-chakra-ui": "1.0.0-beta.2",
     "@types/mapbox-gl": "^2.6.3",
     "@types/react-scroll": "^1.8.3",
     "@types/react-scrollspy": "^3.3.5",
     "esm-ts-node": "^1.0.7",
     "framer-motion": "^4.1.17",
-    "gatsby": "^4.7.2",
+    "gatsby": "^4.9.0",
     "gatsby-plugin-gatsby-cloud": "^4.7.0",
     "gatsby-plugin-image": "^2.7.0",
     "gatsby-plugin-manifest": "^4.7.0",
@@ -92,10 +90,8 @@
     "dotenv": "^10.0.0",
     "gatsby-plugin-compile-es6-packages": "^2.1.1",
     "gatsby-plugin-emotion": "^6.13.0",
-    "gatsby-plugin-ts-config": "^2.1.3",
     "gh-pages": "^3.2.3",
     "prettier": "2.3.2",
-    "storybook-addon-gatsby": "^0.0.5",
     "typescript": "^4.4.2"
   },
   "resolutions": {
